lab3/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
tensorflow==2.16.0
numpy==2.0.0